refactor LoginStateListener
